finished moving style functions to styler
